Fixed bug where subject questions were not being rendered outside admin
