+++ src/app/article-creator/_components/custom_questions/QuestionInstance.tsx
@@ -88,6 +88,7 @@ export const QuestionsOutput: React.FC<{ instanceId: string }> = ({
   instanceId,
 }) => {
   const { questions } = useSyncedQuestions(instanceId);
+  console.log("Questions:", questions);
 
   return (
     <div className="mt-8">
