+++ src/app/article-creator/_components/custom_questions/QuestionsAddCard.tsx
@@ -33,68 +33,10 @@ export class QuestionsAddCard implements BlockToolConstructable {
     };
   }
 
-  // Function to repropagate the questions with parsed data and file URLs
-  private async repropagateQuestions() {
-    const key = getKey();
-    const user = await getUser();  
-
-    if(user && (user.access !== "admin" && user.access !== "member")) {
-      console.log("User is not authorized to perform this action.");
-      return;
-    }
-
-    const storageKey = `questions_${this.instanceId}`;
-    
-    try {
-      // Load questions from Firestore, if available
-      const docRef = doc(db, "pages", key);  // reference needs to be from subject...how do I get that?
-      console.log("docRef:", docRef);
-      const docSnap = await getDoc(docRef);
-
-      if (docSnap.exists()) {
-        const data = docSnap.data().data.blocks;
-        data.forEach((block: any) => {
-          if (block.type === "questionsAddCard") {
-            const questionsFromDb: QuestionFormat[] = block.data.questions.map((question: any) => ({
-              ...question,
-              body: question.body || { value: "" },  // Ensure body exists
-              options: question.options.map((option: any) => ({
-                ...option,
-                value: option.value || { value: "" },  // Ensure value exists
-              })),
-              correct: question.correct || [],
-              explanation: question.explanation || { value: "" },
-              course_id: question.course_id || "",
-              unit_ids: question.unit_ids || [],
-              subunit_ids: question.subunit_ids || [],
-            }));
-    
-            // Update local storage
-            localStorage.setItem(storageKey, JSON.stringify(questionsFromDb));
-
-             // Trigger a manual event to notify listeners that localStorage was updated
-             const event = new Event("questionsUpdated");
-             window.dispatchEvent(event);
-          }
-        });
-      } else {
-        console.log("No document found for questions");
-        return this.data.questions || []; 
-      }
-    } catch (error) {
-      console.error("Error fetching questions from Firestore:", error);
-      return [];
-    }
-  }
-
   render(): HTMLElement {
     const wrapper = document.createElement("div");
     wrapper.classList.add("custom-question-tool");
     const root = ReactDOM.createRoot(wrapper);
-    // Repropagate questions and set the state
-    this.repropagateQuestions().then((questionsFromDb) => {
-      this.questions = questionsFromDb;
-    });
 
     root.render(<QuestionsInput instanceId={this.instanceId} />);
     return wrapper;
