+++ src/app/subject/[slug]/[unit]/[articleNumber]/page.tsx
@@ -8,15 +8,19 @@ import SubjectSidebar from "@/components/subjectHomepage/subject-sidebar";
 import TableOfContents from "@/components/subjectHomepage/table-of-contents";
 import { useEffect, useState } from "react";
 import { doc, getDoc } from "firebase/firestore";
-import Renderer from "@/app/article-creator/_components/Renderer";
 import { Content } from "@/types/content";
 import { useUser } from "@/components/hooks/UserContext";
+import {
+  getKey,
+  revertTableObjectToArray,
+} from "@/app/article-creator/_components/ArticleCreator";
+import { OutputData } from "@editorjs/editorjs";
+import Renderer from "@/app/article-creator/_components/Renderer";
 
 const Page = ({ params }: { params: { slug: string } }) => {
   const [subject, setSubject] = useState<Subject | null>(null);
   const [content, setContent] = useState<Content | null>(null);
   const { user, loading, error, setError, setLoading } = useUser();
-  
 
   const pathParts = window.location.pathname.split("/").slice(-2);
   const formattedTitle =
@@ -32,7 +36,7 @@ const Page = ({ params }: { params: { slug: string } }) => {
           // Reference to the document in Firestore using the slug
           const subjectDocRef = doc(db, "subjects", params.slug);
           const subjectDocSnap = await getDoc(subjectDocRef);
-          
+
           if (subjectDocSnap.exists()) {
             // Convert Firestore document data to Subject type
             setSubject(subjectDocSnap.data() as Subject);
@@ -40,11 +44,16 @@ const Page = ({ params }: { params: { slug: string } }) => {
             setError("Subject not found. That's probably us, not you.");
           }
 
-          const pathParts = window.location.pathname.split("/").slice(-3);
-          const pageDocRef = doc(db, "pages", pathParts.join("-"));
+          const key = getKey();
+          const pageDocRef = doc(db, "pages", key);
           const pageDocSnap = await getDoc(pageDocRef);
+
           if (pageDocSnap.exists()) {
+            const data = pageDocSnap.data()?.data as OutputData;
+            revertTableObjectToArray(data);
+            pageDocSnap.data()!.data = data;
             setContent(pageDocSnap.data() as Content);
+            console.log("Content:", pageDocSnap.data());
           } else {
             setError("Content not found. That's probably us, not you.");
           }
